Split Core/NoViewModel base fragments so updated fragments can use the base item(s) but not need to implement a view model just yet (to speed up re-skinning)
