+++ app/src/main/java/arcus/app/common/fragments/CoreFragment.kt
@@ -16,79 +16,32 @@
 package arcus.app.common.fragments
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.Menu
-import android.view.MenuInflater
 import android.view.View
-import android.view.ViewGroup
-import android.view.ViewStub
-import androidx.annotation.LayoutRes
-import androidx.annotation.MenuRes
 import androidx.core.view.isGone
 import androidx.core.view.isVisible
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import arcus.app.R
 import arcus.presentation.common.view.ViewState
 import arcus.presentation.common.view.ViewStateViewModel
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 
-abstract class CoreFragment<T : ViewStateViewModel<*>> : Fragment() {
-    protected val logger: Logger = LoggerFactory.getLogger(this.javaClass)
-
+/**
+ * Used for classes that have been updated to use a view model. Once all classes are updated  to have a ViewModel this
+ * should have everything from [NoViewModelFragment] consolidated into this [CoreFragment] and [NoViewModelFragment]
+ * removed.
+ */
+abstract class CoreFragment<T : ViewStateViewModel<*>> : NoViewModelFragment() {
     /**
      * The fragments [viewModel].
      */
     protected lateinit var viewModel: T
 
     /**
-     * A container holding a progress spinner (indeterminate) that can be shown while an action
-     * is pending.
-     */
-    protected lateinit var progressContainer: View
-
-    /**
      * The view model class this fragment should load.
      */
     abstract val viewModelClass: Class<T>
 
-    /**
-     * The title of this fragment that should be set on an action bar (if applicable).
-     */
-    abstract val title: String
-
-    /**
-     * The layout to be inflated in this Fragment.
-     */
-    abstract val layoutId: Int
-        @LayoutRes get
-
-    /**
-     * The menu to inflate, if any.
-     */
-    open val menuId: Int? = null
-        @MenuRes get
-
-    override fun onCreateView(
-            inflater: LayoutInflater,
-            container: ViewGroup?,
-            savedInstanceState: Bundle?
-    ): View? {
-        val view = inflater.inflate(R.layout.fragment_core, container, false)
-        view?.findViewById<ViewStub>(R.id.coreViewContainer)?.let { stub ->
-            stub.layoutResource = layoutId
-            stub.inflate()
-        }
-
-        setHasOptionsMenu(true)
-        return view
-    }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        progressContainer = view.findViewById(R.id.progressContainer)
         viewModel = ViewModelProviders.of(this).get(viewModelClass)
         viewModel.viewState.observe(viewLifecycleOwner, Observer {
             when (it) {
@@ -96,25 +49,5 @@ abstract class CoreFragment<T : ViewStateViewModel<*>> : Fragment() {
                 else -> progressContainer.isGone = true
             }
         })
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
-        menuId?.let { id -> inflater.inflate(id, menu) }
-    }
-
-    override fun onResume() {
-        super.onResume()
-        setTitle()
-    }
-
-    /**
-     * Call to set the title of the view to the [title].
-     */
-    protected fun setTitle() {
-        requireActivity().let {
-            it.title = title
-            it.invalidateOptionsMenu()
-        }
     }
 }
